move react components into components folder
